refactored, making HudMasterClientOptions more independent
