+++ time-arena-game/Assets/Scripts/Hud/HudMasterClientOptions.cs
@@ -9,30 +9,31 @@ public class HudMasterClientOptions : MonoBehaviour
 {
     [SerializeField] private GameObject _masterClientOptions;
     [SerializeField] private Text _text;
-    private bool _isCountingTillGameStart;
-    private float _secondsTillGame;
+    private PreGameController _preGame;
+
+    void Start()
+    {
+        _preGame = FindObjectOfType<PreGameController>();
+        if (_preGame == null) Debug.LogError("PreGameController not found");
+    }
 
     private void LateUpdate()
     {
         // If master client, show options message.
         _masterClientOptions.SetActive(SceneManager.GetActiveScene().name == "PreGameScene");
 
-        if (_isCountingTillGameStart)
+        if (_preGame != null && _preGame.IsCountingDown())
         {
-            var timeLeft = System.Math.Round(_secondsTillGame, 0);
+            float secondsTillGame = _preGame.GetSecondsTillGame();
+            var timeLeft = System.Math.Round(secondsTillGame, 0);
             _text.text = $"Starting in {timeLeft}s";
-            if (System.Math.Round(_secondsTillGame, 0) <= 0.0f)
+            if (System.Math.Round(secondsTillGame, 0) <= 0.0f)
             {
                 _text.text = "Loading...";
             }
-        } else {
-          _text.text = "Press F to Start";
         }
+        else _text.text = "Press F to Start";
     }
 
     public void SetActive(bool value) { _masterClientOptions.SetActive(value); }
-
-    public void SetIsCountingTillStart(bool value) { _isCountingTillGameStart = value; }
-
-    public void SetSecondsTillGame(float seconds) { _secondsTillGame = seconds; }
 }
